requirements.txt
# requirements.txt
Django>=5.0
djangorestframework
geopandas
shapely
